feat(focus-trap): localize focus trap package and uninstall the previous one
The previous package contained lodash and also some operations that aren't necessary for our usecase.

Hence for performance reasons, it is better for us to localize the package.

We will also need the package not just for modals, but also other elements such the the button menu in the navigation bar.